+++ pdfs/src/Bs2Jpsifzero_Signal_v5_forcombined.cpp
@@ -1058,7 +1058,7 @@ double Bs2Jpsifzero_Signal_v5_forcombined::diffXsecCompositeNorm1( int resolutio
 
 	for( unsigned int islice = 0; islice < (unsigned) timeAcc->numberOfSlices(); ++islice )
 	{
-		timeBinNum = islice;
+		timeBinNum = (int)islice;
 		//De cache the time integrals  (unles using event Resolution
 		/*if( ! useEventResolution() )*/ this->deCacheTimeIntegrals( (unsigned)resolutionIndex, islice ) ;
